Add byoc domain names test cases
